+++ pytorch_forecasting/layers/_blocks/__init__.py
@@ -1,3 +1,4 @@
 from pytorch_forecasting.layers._blocks._residual_block_dsipts import ResidualBlock
+from pytorch_forecasting.layers._blocks._transformer_block import _TransformerBlock
 
-__all__ = ["ResidualBlock"]
+__all__ = ["ResidualBlock", "_TransformerBlock"]
