+++ pytorch_forecasting/layers/_blocks/_transformer_block.py
@@ -1,0 +1,45 @@
+"""
+Pre-norm Transformer Encoder Block for PTF.
+"""
+
+import torch
+import torch.nn as nn
+
+
+class _TransformerBlock(nn.Module):
+    """
+    Pre-norm transformer encoder block (MHSA + FFN).
+
+    Parameters
+    ----------
+    d_model : int
+        Model dimension.
+    n_heads : int
+        Number of attention heads.
+    d_ff : int
+        Feed-forward hidden dimension.
+    dropout : float
+        Dropout probability.
+    """
+
+    def __init__(self, d_model: int, n_heads: int, d_ff: int, dropout: float = 0.1):
+        super().__init__()
+        self.norm1 = nn.LayerNorm(d_model)
+        self.norm2 = nn.LayerNorm(d_model)
+        self.attn = nn.MultiheadAttention(
+            embed_dim=d_model, num_heads=n_heads, dropout=dropout, batch_first=True
+        )
+        self.ff = nn.Sequential(
+            nn.Linear(d_model, d_ff),
+            nn.GELU(),
+            nn.Dropout(dropout),
+            nn.Linear(d_ff, d_model),
+            nn.Dropout(dropout),
+        )
+
+    def forward(self, x: torch.Tensor) -> torch.Tensor:
+        normed = self.norm1(x)
+        attn_out, _ = self.attn(normed, normed, normed)
+        x = x + attn_out
+        x = x + self.ff(self.norm2(x))
+        return x
