+++ pytorch_forecasting/layers/_embeddings/__init__.py
@@ -6,13 +6,17 @@ from pytorch_forecasting.layers._embeddings._data_embedding import (
     DataEmbedding_inverted,
 )
 from pytorch_forecasting.layers._embeddings._en_embedding import EnEmbedding
+from pytorch_forecasting.layers._embeddings._patch_embedding import _PatchEmbedding
 from pytorch_forecasting.layers._embeddings._positional_embedding import (
     PositionalEmbedding,
+    _PositionalEmbedding,
 )
 from pytorch_forecasting.layers._embeddings._sub_nn import embedding_cat_variables
 
 __all__ = [
     "PositionalEmbedding",
+    "_PositionalEmbedding",
+    "_PatchEmbedding",
     "DataEmbedding_inverted",
     "EnEmbedding",
     "embedding_cat_variables",
